requirements.txt
# Kafka
kafka-python==2.0.2

# LLM Engines  
google-generativeai==0.8.3
requests==2.31.0

# Data Models
pydantic==2.8.2

# CLI
rich==13.7.1
click==8.1.7

# Utilities
python-dotenv==1.0.1
asyncio-throttle==1.0.2

# Logging
structlog==24.2.0

# Language Detection
langdetect==1.0.9
